Fixed race condition
The webcam video capture was stuck in a queue waiting for a frame to come with visualization labels, but since the threads that place the frame in the queue terminated first, the video capture waited forever. Changed the stop function in video capture to put a special STOP flag in the queue for the thread to know when to stop.